changes towards int64 support.

+++ util/raw-functions.hpp
@@ -125,7 +125,7 @@ extern "C" void flushArrayEnd(char* fileName);
 
 extern "C" void flushInt(int toFlush, char* fileName);
 
-extern "C" void flushInt64(int toFlush, char* fileName);
+extern "C" void flushInt64(size_t toFlush, char* fileName);
 
 extern "C" void flushDouble(double toFlush, char* fileName);
